PPU: LY == LYC interrupt
If the index of the current row is equal to the value in the
LYC register (and the STAT condition for LY == LYC is set),
we request the STAT interrupt.

We also set bit 2 of the STAT register, if the two are equal,
regardless of whether this will trigger an interrupt or not.

This required a change in two unrelated tests, because setting
the LY to 0 without setting LYC to something else will
automatically trigger the condition. However, I wanted the
test cases to be independent of the LY == LYC condition,
so I set the LY to 1 instead.